doc: improve the doc of the config module
