+++ src/vsmtp/vsmtp-common/src/lib.rs
@@ -39,6 +39,20 @@ pub const SUBMISSION_PORT: u16 = 587;
 /// Defined in [RFC8314](https://tools.ietf.org/html/rfc8314)
 pub const SUBMISSIONS_PORT: u16 = 465;
 
+/// Type of SMTP connection.
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Hash, strum::Display)]
+pub enum ConnectionKind {
+    /// Connection coming for relay (MTA on port 25)
+    /// see <https://datatracker.ietf.org/doc/html/rfc5321>
+    Relay,
+    /// Connection coming for submission (MSA on port 587)
+    /// see <https://datatracker.ietf.org/doc/html/rfc6409>
+    Submission,
+    /// Connection coming for submissionS (MSA on port 465)
+    /// see <https://datatracker.ietf.org/doc/html/rfc8314>
+    Tunneled,
+}
+
 mod log_channels {
     pub const QUEUE: &str = "server::queue";
 }
