+++ src/vsmtp/vsmtp-config/src/parser/tls_private_key.rs
@@ -88,10 +88,9 @@ S: serde::Serializer,
 mod tests {
     use std::io::Write;
 
+    use crate::field::TlsFile;
     use vsmtp_common::re::serde_json;
     use vsmtp_test::get_tls_file;
-
-    use crate::TlsFile;
 
     #[derive(Debug, serde::Serialize, serde::Deserialize)]
     struct S {
