+++ src/vsmtp/vsmtp-server/src/lib.rs
@@ -37,7 +37,7 @@ pub use receiver::MailHandler;
 /// SMTP auth extension implementation
 pub mod auth;
 pub use channel_message::ProcessMessage;
-pub use receiver::{handle_connection, AbstractIO, Connection, ConnectionKind, OnMail};
+pub use receiver::{handle_connection, AbstractIO, Connection, OnMail};
 pub use runtime::start_runtime;
 pub use server::{socket_bind_anyhow, Server};
