+++ src/vsmtp/vsmtp-test/src/receiver.rs
@@ -18,11 +18,11 @@
 use vsmtp_common::{
     mail_context::MessageBody,
     re::{anyhow, tokio},
-    CodeID,
+    CodeID, ConnectionKind,
 };
 use vsmtp_config::Config;
 use vsmtp_rule_engine::rule_engine::RuleEngine;
-use vsmtp_server::{auth, handle_connection, Connection, ConnectionKind, OnMail};
+use vsmtp_server::{auth, handle_connection, Connection, OnMail};
 
 /// A type implementing Write+Read to emulate sockets
 pub struct Mock<'a, T: AsRef<[u8]> + Unpin> {
