+++ src/vsmtp/vsmtp-test/src/tests/tls/starttls.rs
@@ -17,7 +17,7 @@
 use super::get_tls_config;
 use crate::{test_receiver, tests::tls::test_starttls};
 use vsmtp_common::re::{anyhow, tokio};
-use vsmtp_config::TlsSecurityLevel;
+use vsmtp_config::field::TlsSecurityLevel;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 3)]
 async fn simple() {
